Group by search

+++ backend/src/main/java/H/dsvendas/dto/SaleDTO.java
@@ -7,8 +7,8 @@ import H.Dsvendas.entities.Sale;
 public class SaleDTO {
 	
 	private Long id;
-	private Integer visited;
-	private Integer deals;
+	private Long visited;
+	private Long deals;
 	private Double amount;
 	private LocalDate date;
 	
@@ -17,7 +17,7 @@ public class SaleDTO {
 	public SaleDTO() {
 	}
 
-	public SaleDTO(Long id, Integer visited, Integer deals, Double amount, LocalDate date, SellerDTO seller) {
+	public SaleDTO(Long id, Long visited, Long deals, Double amount, LocalDate date, SellerDTO seller) {
 		this.id = id;
 		this.visited = visited;
 		this.deals = deals;
@@ -43,19 +43,19 @@ public class SaleDTO {
 		this.id = id;
 	}
 
-	public Integer getVisited() {
+	public Long getVisited() {
 		return visited;
 	}
 
-	public void setVisited(Integer visited) {
+	public void setVisited(Long visited) {
 		this.visited = visited;
 	}
 
-	public Integer getDeals() {
+	public Long getDeals() {
 		return deals;
 	}
 
-	public void setDeals(Integer deals) {
+	public void setDeals(Long deals) {
 		this.deals = deals;
 	}
